app/requirements.txt
django
prettyconf
Pillow
requests
django-debug-toolbar
django-bootstrap-v5
WeasyPrint
django-rosetta
djangorestframework